tasks/prepare_pwd_parcels/requirements.txt
functions-framework==3.*
google-cloud-storage==2.*
python-dotenv==1.*
pandas==2.*
geopandas==0.14.*
pyarrow==15.*
shapely==2.*
pyproj==3.*
